% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_kinetics.R
\docType{data}
\name{equations}
\alias{equations}
\title{Use kinetic parameters to simulate data over time, substrate or rate}
\format{
An object of class \code{list} of length 3.
}
\usage{
equations
}
\arguments{
\item{time}{A seq(a,b,n) vector setting size of simulation}

\item{params}{list of parameters useful for the simulation (S0, Km, Vmax, Ki, Ka)}
}
\value{
df dataset with a time, substrate or reaction rate structure
}
\description{
Use kinetic parameters to simulate data over time, substrate or rate
}
\examples{
simulate_kinetics(time, params=list(S0=1,Km=0.0149,Vmax=2.533), inhibition=NULL)
}
\keyword{datasets}
