% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_params.R
\name{fit_params}
\alias{fit_params}
\title{Determination of Vmax and Km parameters for irreversible MM equation}
\usage{
fit_params(
  df,
  model = c("one-substrate", "pH", "temperature", "ionic", "inhibition", "allosteric"),
  method = c("lineweaver", "eadie", "hanes", "nonlinear"),
  level = 0.95,
  ...
)
}
\arguments{
\item{df}{Substrate-time or substrate-reaction rate dataset}

\item{model}{Selection of the model type ("time" for substrate-time or "rate" for substrate-reaction rate)}

\item{method}{Selection of a linearization strategy ("lb" for Lineweaver-Burk, "eh" for Eadie-Hofstee, "hw" for Hanes-Woolf)}
}
\value{
A list with Km and Vmax parameters
}
\description{
Determination of Vmax and Km parameters for irreversible MM equation
}
\examples{
fit_params(df, model="time", method="lin")
}
